Rename template tag library and add warning for the old name
