+++ menu/__init__.py
@@ -5,5 +5,6 @@ from menu import *
 warnings.warn(
 	"Importing from 'menu' will be removed in django-simple-menu v2.3 and/or "
 	"v3.0. Use imports from 'simple_menu' package instead.",
-	DeprecationWarning
+	DeprecationWarning,
+	stacklevel=2
 )
